pyproject.toml
[project]
name = "mind"
version = "0.1.0"
description = ""
authors = [{ name = "Lorena Calvo-Bartolomé & Alonso Madroñal de Mesa" }]
requires-python = "==3.12.*"
readme = "README.md"
license = {text = "MIT"}

dependencies = [
    "colorama==0.4.6",
    "fasttext==0.9.3",
    "datasets==4.0.0",
    "faiss-cpu",
    "huggingface-hub==0.34.4",
    "joblib==1.4.2",
    "kneed==0.8.5",
    "nltk==3.9.1",
    "numpy<2.0.0,>=1.18.5",
    "ollama==0.5.3",
    "openai==1.106.1",
    "pandas==2.3.2",
    "protobuf>=4.25.8",
    "pyarrow==16.1.0",
    "pyLDAvis==3.4.1",
    "python-dotenv==1.1.1",
    "PyYAML==6.0.2",
    "Requests==2.32.5",
    "scikit-learn==1.7.1",
    "sentence-transformers==5.1.0",
    "torch==2.8.0",
    "tqdm==4.67.1",
    "transformers==4.56.0",
    "sentencepiece==0.2.1",
    "scikit_posthocs",
    "openpyxl",
    "krippendorff",
    "irrCAC",
    "Wikipedia-API",
    "dask",
    "langdetect",
    "py7zr",
]

[build-system]
requires = ["hatchling>=1.26", "pip>=24"]
build-backend = "hatchling.build"

[tool.hatch.build]
sources = ["src"]

[tool.uv]
package = false

[tool.hatch.metadata]
allow-direct-references = true